Creating logic to send or remove pending reactions is in process.

+++ SceytChatUiKit/src/main/java/com/sceyt/sceytchatuikit/data/models/messages/ReactionData.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class ReactionData(
         var key: String,
-        val score: Long = 1,
+        val score: Int = 1,
         val containsSelf: Boolean = false) : Parcelable {
 
     override fun equals(other: Any?): Boolean {
